Update README with app status and fix instructions numbering
Added a note on README to inform users about the current state of the application and reassure them about the progress of its development. This is necessary to manage expectations among users. Also, fixed numbering in the section describing how to contribute to the project. The numbers were inaccurate due to a previous omission of steps.