Add PyCustomToggleButton.
PyCustomToggleButton inherits from QPushButton and creates two custom signals, button_checked, and button_unchecked, and enforces a limit of at most 2 buttons checked, it will uncheck the least recently checked button when trying to check a button if checked_buttons >= 2.

The custom signals emit the name of the button with a '_' prepended and _btn removed from the end ex: crusader_btn would emit _crusader. This is used by the set_influence and remove_influence methods in the GUIUpdater which adds them to a list of influences applied to an item