Add splash screen to ExileCraft
A splash screen was added to the ExileCraft application. This screen is the initial interface the user sees when the program is started. It displays a progress bar that fills as the program loads with messages indicating that the program is loading, and finally it shows the main window of the application upon reaching 100%.